DesplegarApp/requirements.txt

flask
python-dotenv
gunicorn